% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getben.R
\name{getben}
\alias{getben}
\title{Download das informações consolidadas de beneficiários da ANS}
\usage{
getben(ano, mes, cobertura, tipo_contrato, autogestao = FALSE, uf = "all")
}
\arguments{
\item{ano.}{Ano de interesse no formato "aaaa".}

\item{mes.}{Mês de interesse no formato "mm".}

\item{cobertura.}{Tipo de cobertura de plano do beneficiário - "Médico-hospitalar" ou "Odontológico".}

\item{tipo_contrato.}{Tipo de contratação do plano do beneficiário - "Empresarial", "Adesão", "Individual".}

\item{autogestao.}{Parâmetro para exclusão das operadoras que seguem a modalidade de autogestão. Use autogestao = TRUE para incluir essa modalidade.}

\item{uf.}{UFs de interesse.}
}
\description{
Executa o download das informações consolidadas de beneficiários da ANS para o período e UFs de interesse
}
\examples{

getben(ano = "2020", mes = "12", cobertura = "Médico-hospitalar", tipo_contrato = "Individual", uf = "SP")

getben(ano = "2020", mes = "12", cobertura = "Odontológico", tipo_contrato = "Empresarial", uf = c("MG","SP","RJ"))

}
